refactor: extract business logic from main.rs into library modules
Move git operations, discovery orchestration, test execution, graph
rendering, and watch mode out of the binary's main.rs into dedicated
library modules (git.rs, discovery.rs, execution.rs, graph.rs, watch.rs).
This makes the logic testable without binary context, reusable by other
tools, and reduces main.rs from ~1750 lines to ~170 lines of CLI glue.

All tests moved with their corresponding functions.
