+++ crates/tryke/src/lib.rs
@@ -1,2 +1,7 @@
 pub mod cli;
 pub mod cli_docs;
+pub mod discovery;
+pub mod execution;
+pub mod git;
+pub mod graph;
+pub mod watch;
