+++ crates/tryke/src/main.rs
@@ -1,457 +1,19 @@
-use std::{
-    collections::BTreeSet,
-    env,
-    path::{Path, PathBuf},
-    time::{Duration, Instant},
-};
+use std::{env, time::Instant};
 
 use anyhow::Result;
 use clap::Parser;
-use log::{debug, warn};
-use tokio_stream::StreamExt;
+use log::debug;
 use tryke::cli::{Cli, Commands, ReporterFormat};
-use tryke_config::load_effective_config;
-use tryke_discovery::Discoverer;
+use tryke::discovery::{discover_tests, discover_tests_changed_first, resolved_excludes};
+use tryke::execution::run_tests;
+use tryke::graph::run_graph;
+use tryke::watch::run_watch;
 use tryke_reporter::{
     DotReporter, JSONReporter, JUnitReporter, LlmReporter, ProgressReporter, Reporter,
     TextReporter, Verbosity,
 };
-use tryke_runner::{WorkerPool, check_python_version, resolve_python};
+use tryke_types::ChangedSelectionSummary;
 use tryke_types::filter::TestFilter;
-use tryke_types::{ChangedSelectionSummary, RunSummary, TestOutcome};
-
-fn worker_pool_size() -> usize {
-    std::thread::available_parallelism().map_or(4, std::num::NonZero::get)
-}
-
-struct DiscoverySelection {
-    tests: Vec<tryke_types::TestItem>,
-    changed_files: Option<usize>,
-    /// In changed-first mode, how many tests at the front are "changed" tests.
-    #[cfg_attr(not(test), allow(dead_code))]
-    changed_prefix_len: Option<usize>,
-}
-
-fn resolved_excludes(root: &Path, cli_excludes: &[String], cli_includes: &[String]) -> Vec<String> {
-    if !cli_excludes.is_empty() {
-        return cli_excludes.to_vec();
-    }
-    let includes = cli_includes
-        .iter()
-        .collect::<std::collections::HashSet<_>>();
-    load_effective_config(root)
-        .discovery
-        .exclude
-        .into_iter()
-        .filter(|exclude| !includes.contains(exclude))
-        .collect()
-}
-
-/// Discover tests, optionally restricting to changed files.
-fn discover_tests(
-    root: &Path,
-    changed: bool,
-    base_branch: Option<&str>,
-    excludes: &[String],
-) -> DiscoverySelection {
-    if changed {
-        let mut discoverer = Discoverer::new_with_excludes(root, excludes);
-        discoverer.rediscover();
-        match resolve_changed_files(root, base_branch) {
-            Some(changed_files) if !changed_files.is_empty() => {
-                debug!("--changed: {} git-changed files", changed_files.len());
-                DiscoverySelection {
-                    tests: discoverer.tests_for_changed(&changed_files),
-                    changed_files: Some(changed_files.len()),
-                    changed_prefix_len: None,
-                }
-            }
-            Some(_) => {
-                warn!("--changed: no changed files found via git, running all tests");
-                DiscoverySelection {
-                    tests: discoverer.tests(),
-                    changed_files: None,
-                    changed_prefix_len: None,
-                }
-            }
-            None => {
-                warn!("--changed: git unavailable or failed, running all tests");
-                DiscoverySelection {
-                    tests: discoverer.tests(),
-                    changed_files: None,
-                    changed_prefix_len: None,
-                }
-            }
-        }
-    } else {
-        DiscoverySelection {
-            tests: tryke_discovery::discover_from_with_excludes(root, excludes),
-            changed_files: None,
-            changed_prefix_len: None,
-        }
-    }
-}
-
-/// Discover all tests but place changed tests first in the returned list.
-fn discover_tests_changed_first(
-    root: &Path,
-    base_branch: Option<&str>,
-    excludes: &[String],
-) -> DiscoverySelection {
-    let mut discoverer = Discoverer::new_with_excludes(root, excludes);
-    discoverer.rediscover();
-    let changed_files = resolve_changed_files(root, base_branch);
-    let all_tests = discoverer.tests();
-    match changed_files {
-        Some(cf) if !cf.is_empty() => {
-            let changed_tests = discoverer.tests_for_changed(&cf);
-            let changed_ids: std::collections::HashSet<String> =
-                changed_tests.iter().map(|t| t.id()).collect();
-            let (first, rest): (Vec<_>, Vec<_>) = all_tests
-                .into_iter()
-                .partition(|t| changed_ids.contains(&t.id()));
-            let changed_prefix_len = first.len();
-            let mut tests = first;
-            tests.extend(rest);
-            DiscoverySelection {
-                tests,
-                changed_files: Some(cf.len()),
-                changed_prefix_len: Some(changed_prefix_len),
-            }
-        }
-        Some(_) => {
-            warn!("--changed-first: no changed files found, running all tests in default order");
-            DiscoverySelection {
-                tests: all_tests,
-                changed_files: None,
-                changed_prefix_len: None,
-            }
-        }
-        None => {
-            warn!("--changed-first: git unavailable, running all tests in default order");
-            DiscoverySelection {
-                tests: all_tests,
-                changed_files: None,
-                changed_prefix_len: None,
-            }
-        }
-    }
-}
-
-async fn run_tests(
-    reporter: &mut dyn Reporter,
-    root: &Path,
-    tests: Vec<tryke_types::TestItem>,
-    maxfail: Option<usize>,
-    workers: Option<usize>,
-    discovery_duration: Option<Duration>,
-    changed_selection: Option<ChangedSelectionSummary>,
-) -> Result<()> {
-    let python = resolve_python(root);
-    check_python_version(&python, root)?;
-    let pool_size = workers.unwrap_or_else(|| tests.len().min(worker_pool_size()));
-    let pool = WorkerPool::new(pool_size, &python, root);
-    pool.warm().await;
-    report_cycle(
-        reporter,
-        tests,
-        &pool,
-        maxfail,
-        discovery_duration,
-        changed_selection,
-    )
-    .await?;
-    pool.shutdown();
-    Ok(())
-}
-
-async fn report_cycle(
-    reporter: &mut dyn Reporter,
-    tests: Vec<tryke_types::TestItem>,
-    pool: &WorkerPool,
-    maxfail: Option<usize>,
-    discovery_duration: Option<Duration>,
-    changed_selection: Option<ChangedSelectionSummary>,
-) -> Result<()> {
-    use std::collections::HashSet;
-
-    let file_count = tests
-        .iter()
-        .filter_map(|t| t.file_path.as_ref())
-        .collect::<HashSet<_>>()
-        .len();
-
-    let start_time = chrono::Local::now().format("%H:%M:%S").to_string();
-
-    let start = Instant::now();
-    reporter.on_run_start(&tests);
-
-    let mut passed = 0usize;
-    let mut failed = 0usize;
-    let mut skipped = 0usize;
-    let mut errors = 0usize;
-    let mut xfailed = 0usize;
-    let mut todo = 0usize;
-
-    // Short-circuit skip/todo tests — no worker needed
-    let (run_tests, shortcircuit): (Vec<_>, Vec<_>) = tests
-        .into_iter()
-        .partition(|t| t.skip.is_none() && t.todo.is_none());
-
-    for t in shortcircuit {
-        let outcome = if t.todo.is_some() {
-            todo += 1;
-            TestOutcome::Todo {
-                description: t.todo.clone(),
-            }
-        } else {
-            skipped += 1;
-            TestOutcome::Skipped {
-                reason: t.skip.clone(),
-            }
-        };
-        let result = tryke_types::TestResult {
-            test: t,
-            outcome,
-            duration: std::time::Duration::ZERO,
-            stdout: String::new(),
-            stderr: String::new(),
-        };
-        reporter.on_test_complete(&result);
-    }
-
-    let mut stream = pool.run(run_tests);
-    while let Some(result) = stream.next().await {
-        match &result.outcome {
-            TestOutcome::Passed => passed += 1,
-            TestOutcome::Failed { .. } | TestOutcome::XPassed => failed += 1,
-            TestOutcome::Skipped { .. } => skipped += 1,
-            TestOutcome::Error { .. } => errors += 1,
-            TestOutcome::XFailed { .. } => xfailed += 1,
-            TestOutcome::Todo { .. } => todo += 1,
-        }
-        reporter.on_test_complete(&result);
-        if let Some(max) = maxfail
-            && failed >= max
-        {
-            break;
-        }
-    }
-
-    reporter.on_run_complete(&RunSummary {
-        passed,
-        failed,
-        skipped,
-        errors,
-        xfailed,
-        todo,
-        duration: discovery_duration.unwrap_or_default() + start.elapsed(),
-        discovery_duration,
-        test_duration: Some(start.elapsed()),
-        file_count,
-        start_time: Some(start_time),
-        changed_selection,
-    });
-
-    Ok(())
-}
-
-fn clear_if_tty() {
-    use std::io::IsTerminal;
-    if std::io::stdout().is_terminal() {
-        let _ = clearscreen::clear();
-    }
-}
-
-async fn run_watch(
-    reporter: &mut dyn Reporter,
-    root: Option<&Path>,
-    excludes: &[String],
-    test_filter: &TestFilter,
-    maxfail: Option<usize>,
-    workers: Option<usize>,
-) -> Result<()> {
-    let cwd = env::current_dir()?;
-    let root = root.unwrap_or(&cwd);
-    let mut discoverer = Discoverer::new_with_excludes(root, excludes);
-
-    let python = resolve_python(root);
-    check_python_version(&python, root)?;
-    let pool_size = workers.unwrap_or_else(worker_pool_size);
-    let pool = WorkerPool::new(pool_size, &python, root);
-    pool.warm().await;
-
-    clear_if_tty();
-    let disc_start = Instant::now();
-    let tests = test_filter.apply(discoverer.rediscover());
-    let disc_dur = Some(disc_start.elapsed());
-    report_cycle(reporter, tests, &pool, maxfail, disc_dur, None).await?;
-
-    let (tx, rx) = std::sync::mpsc::channel::<Vec<PathBuf>>();
-    let _debouncer = tryke_server::watcher::spawn_watcher(root, excludes, tx)?;
-
-    for paths in &rx {
-        let modules = discoverer.affected_modules(&paths);
-        if !modules.is_empty() {
-            pool.reload(modules).await;
-        }
-        discoverer.rediscover_changed(&paths);
-        clear_if_tty();
-        let disc_start = Instant::now();
-        let tests = test_filter.apply(discoverer.tests_for_changed(&paths));
-        let disc_dur = Some(disc_start.elapsed());
-        report_cycle(reporter, tests, &pool, maxfail, disc_dur, None).await?;
-    }
-
-    pool.shutdown();
-    Ok(())
-}
-
-fn git_paths(root: &Path, args: &[&str]) -> Option<Vec<PathBuf>> {
-    let output = std::process::Command::new("git")
-        .args(args)
-        .current_dir(root)
-        .output()
-        .ok()?;
-    if !output.status.success() {
-        return None;
-    }
-    let text = String::from_utf8(output.stdout).ok()?;
-    let paths: Vec<PathBuf> = text
-        .lines()
-        .filter(|l| !l.is_empty())
-        .map(|l| root.join(l))
-        .collect();
-    Some(paths)
-}
-
-/// Collect changed files from git relative to `root`.
-/// Includes tracked changes since HEAD and untracked files.
-/// Returns `None` if git is unavailable or a command fails.
-fn git_changed_files(root: &Path) -> Option<Vec<PathBuf>> {
-    let tracked = git_paths(root, &["diff", "--name-only", "HEAD"])?;
-    let untracked = git_paths(root, &["ls-files", "--others", "--exclude-standard"])?;
-    let mut paths: Vec<PathBuf> = tracked
-        .into_iter()
-        .chain(untracked)
-        .collect::<BTreeSet<_>>()
-        .into_iter()
-        .collect();
-    paths.sort();
-    Some(paths)
-}
-
-/// Collect files changed on the current branch relative to `base`.
-/// Uses three-dot merge-base diff so only the branch's own changes appear.
-/// Also includes untracked files (not captured by the diff).
-/// Returns `None` if git is unavailable or a command fails.
-fn git_branch_changed_files(root: &Path, base: &str) -> Option<Vec<PathBuf>> {
-    let diff_spec = format!("{base}...HEAD");
-    let branch_diff = git_paths(root, &["diff", "--name-only", &diff_spec])?;
-    let untracked = git_paths(root, &["ls-files", "--others", "--exclude-standard"])?;
-    let mut paths: Vec<PathBuf> = branch_diff
-        .into_iter()
-        .chain(untracked)
-        .collect::<BTreeSet<_>>()
-        .into_iter()
-        .collect();
-    paths.sort();
-    Some(paths)
-}
-
-/// Resolve changed files using either branch mode or HEAD mode.
-fn resolve_changed_files(root: &Path, base_branch: Option<&str>) -> Option<Vec<PathBuf>> {
-    match base_branch {
-        Some(base) => git_branch_changed_files(root, base),
-        None => git_changed_files(root),
-    }
-}
-
-fn run_graph(
-    root: Option<&Path>,
-    excludes: &[String],
-    connected_only: bool,
-    changed: bool,
-    base_branch: Option<&str>,
-) -> Result<()> {
-    let cwd = env::current_dir()?;
-    let root_path = root.unwrap_or(&cwd);
-    let mut discoverer = Discoverer::new_with_excludes(root_path, excludes);
-    discoverer.rediscover();
-
-    let changed_files = if changed {
-        match resolve_changed_files(root_path, base_branch) {
-            Some(paths) if !paths.is_empty() => Some(paths),
-            Some(_) => {
-                println!("No git-visible changed files found.");
-                return Ok(());
-            }
-            None => {
-                println!("Git unavailable or failed; cannot compute changed graph.");
-                return Ok(());
-            }
-        }
-    } else {
-        None
-    };
-
-    let affected = changed_files
-        .as_ref()
-        .map(|paths| discoverer.affected_files(paths))
-        .unwrap_or_default();
-    let changed_set = changed_files
-        .as_ref()
-        .map(|paths| {
-            paths
-                .iter()
-                .cloned()
-                .collect::<std::collections::HashSet<_>>()
-        })
-        .unwrap_or_default();
-
-    let summary = discoverer.import_graph_summary();
-    for entry in &summary {
-        let full_path = root_path.join(&entry.file);
-        if changed && !affected.contains(&full_path) {
-            continue;
-        }
-        if connected_only && entry.imports.is_empty() && entry.imported_by.is_empty() {
-            continue;
-        }
-        let label = if changed {
-            if changed_set.contains(&full_path) {
-                " [changed]"
-            } else {
-                " [affected]"
-            }
-        } else {
-            ""
-        };
-        println!("{}{}", entry.file.display(), label);
-        if entry.imports.is_empty() {
-            println!("  imports:     (none)");
-        } else {
-            let names: Vec<String> = entry
-                .imports
-                .iter()
-                .map(|p| p.display().to_string())
-                .collect();
-            println!("  imports:     {}", names.join(", "));
-        }
-        if entry.imported_by.is_empty() {
-            println!("  imported by: (none)");
-        } else {
-            let names: Vec<String> = entry
-                .imported_by
-                .iter()
-                .map(|p| p.display().to_string())
-                .collect();
-            println!("  imported by: {}", names.join(", "));
-        }
-        println!();
-    }
-    Ok(())
-}
 
 fn build_reporter(format: &ReporterFormat, verbosity: Verbosity) -> Box<dyn Reporter> {
     let use_progress = tryke_reporter::progress::supports_progress()
@@ -637,14 +199,6 @@ mod tests {
 
     use super::*;
 
-    fn test_python_bin() -> String {
-        let root = std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-            .join("../..")
-            .canonicalize()
-            .expect("workspace root");
-        tryke_runner::resolve_python(&root)
-    }
-
     fn group_tests_by_file(tests: Vec<TestItem>) -> Vec<(Option<PathBuf>, Vec<TestItem>)> {
         let mut index: std::collections::HashMap<Option<PathBuf>, usize> =
             std::collections::HashMap::new();
@@ -659,70 +213,6 @@ mod tests {
             }
         }
         groups
-    }
-
-    fn cwd() -> PathBuf {
-        env::current_dir().unwrap_or_else(|_| PathBuf::from("."))
-    }
-
-    async fn run_cycle(
-        reporter: &mut dyn Reporter,
-        discoverer: &mut Discoverer,
-        pool: &WorkerPool,
-    ) -> Result<()> {
-        report_cycle(reporter, discoverer.rediscover(), pool, None, None, None).await
-    }
-
-    #[tokio::test]
-    async fn test_command_text() {
-        let mut reporter = TextReporter::with_writer(Vec::new());
-        let root = cwd();
-        let excludes = resolved_excludes(&root, &[], &[]);
-        let tests = discover_tests(&root, false, None, &excludes).tests;
-        assert!(
-            run_tests(&mut reporter, &root, tests, None, None, None, None)
-                .await
-                .is_ok()
-        );
-    }
-
-    #[tokio::test]
-    async fn test_command_json() {
-        let mut reporter = JSONReporter::with_writer(Vec::new());
-        let root = cwd();
-        let excludes = resolved_excludes(&root, &[], &[]);
-        let tests = discover_tests(&root, false, None, &excludes).tests;
-        assert!(
-            run_tests(&mut reporter, &root, tests, None, None, None, None)
-                .await
-                .is_ok()
-        );
-    }
-
-    #[tokio::test]
-    async fn test_command_dot() {
-        let mut reporter = DotReporter::with_writer(Vec::new());
-        let root = cwd();
-        let excludes = resolved_excludes(&root, &[], &[]);
-        let tests = discover_tests(&root, false, None, &excludes).tests;
-        assert!(
-            run_tests(&mut reporter, &root, tests, None, None, None, None)
-                .await
-                .is_ok()
-        );
-    }
-
-    #[tokio::test]
-    async fn test_command_junit() {
-        let mut reporter = JUnitReporter::with_writer(Vec::new());
-        let root = cwd();
-        let excludes = resolved_excludes(&root, &[], &[]);
-        let tests = discover_tests(&root, false, None, &excludes).tests;
-        assert!(
-            run_tests(&mut reporter, &root, tests, None, None, None, None)
-                .await
-                .is_ok()
-        );
     }
 
     #[test]
@@ -941,48 +431,6 @@ mod tests {
         ));
     }
 
-    #[tokio::test]
-    async fn run_cycle_runs_without_error() {
-        let python_dir = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .join("../../python")
-            .canonicalize()
-            .expect("python/ dir must exist");
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("pyproject.toml"), "").expect("write pyproject.toml");
-        std::fs::write(
-            dir.path().join("test_x.py"),
-            "from tryke import test\n\n@test\ndef test_x(): pass\n",
-        )
-        .expect("write test file");
-        let mut discoverer = Discoverer::new(dir.path());
-        let mut reporter = TextReporter::new();
-        let pool = WorkerPool::with_python_path(
-            1,
-            &test_python_bin(),
-            dir.path(),
-            &[dir.path().to_path_buf(), python_dir],
-        );
-        assert!(
-            run_cycle(&mut reporter, &mut discoverer, &pool)
-                .await
-                .is_ok()
-        );
-    }
-
-    #[tokio::test]
-    async fn run_cycle_with_json_reporter() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("pyproject.toml"), "").expect("write pyproject.toml");
-        let mut discoverer = Discoverer::new(dir.path());
-        let mut reporter = JSONReporter::with_writer(Vec::new());
-        let pool = WorkerPool::new(1, &test_python_bin(), dir.path());
-        assert!(
-            run_cycle(&mut reporter, &mut discoverer, &pool)
-                .await
-                .is_ok()
-        );
-    }
-
     #[test]
     fn test_changed_flag_parsed() {
         let cli = Cli::try_parse_from(["tryke", "test", "--changed"]).unwrap();
@@ -1008,275 +456,6 @@ mod tests {
                 ..
             } if include == &["benchmarks/suites"]
         ));
-    }
-
-    #[test]
-    fn resolved_excludes_reads_pyproject_when_enabled() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(
-            dir.path().join("pyproject.toml"),
-            "[tool.tryke]\nexclude = [\"benchmarks/suites\"]\n",
-        )
-        .expect("write pyproject.toml");
-
-        let excludes = resolved_excludes(dir.path(), &[], &[]);
-        assert_eq!(excludes, vec!["benchmarks/suites"]);
-    }
-
-    #[test]
-    fn resolved_excludes_removes_included_config_excludes() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(
-            dir.path().join("pyproject.toml"),
-            "[tool.tryke]\nexclude = [\"benchmarks/suites\", \"generated\"]\n",
-        )
-        .expect("write pyproject.toml");
-
-        let excludes = resolved_excludes(dir.path(), &[], &["benchmarks/suites".into()]);
-        assert_eq!(excludes, vec!["generated"]);
-    }
-
-    #[test]
-    fn resolved_excludes_prefers_cli_excludes_over_includes() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(
-            dir.path().join("pyproject.toml"),
-            "[tool.tryke]\nexclude = [\"benchmarks/suites\"]\n",
-        )
-        .expect("write pyproject.toml");
-
-        let excludes = resolved_excludes(
-            dir.path(),
-            &["tmp".into(), "cache".into()],
-            &["benchmarks/suites".into()],
-        );
-        assert_eq!(excludes, vec!["tmp", "cache"]);
-    }
-
-    #[test]
-    fn graph_subcommand_parsed() {
-        let cli = Cli::try_parse_from(["tryke", "graph"]).unwrap();
-        assert!(matches!(cli.command, Commands::Graph { .. }));
-    }
-
-    #[test]
-    fn graph_subcommand_changed_parsed() {
-        let cli = Cli::try_parse_from(["tryke", "graph", "--changed"]).unwrap();
-        assert!(matches!(cli.command, Commands::Graph { changed: true, .. }));
-    }
-
-    #[test]
-    fn graph_subcommand_connected_only_parsed() {
-        let cli = Cli::try_parse_from(["tryke", "graph", "--connected-only"]).unwrap();
-        assert!(matches!(
-            cli.command,
-            Commands::Graph {
-                connected_only: true,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn run_graph_prints_entries() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("pyproject.toml"), "").expect("write pyproject.toml");
-        std::fs::write(dir.path().join("utils.py"), "def helper(): pass\n").expect("write");
-        std::fs::write(
-            dir.path().join("test_foo.py"),
-            "from utils import helper\n@test\ndef test_foo(): pass\n",
-        )
-        .expect("write");
-        assert!(run_graph(Some(dir.path()), &[], false, false, None).is_ok());
-    }
-
-    #[test]
-    fn run_graph_connected_only() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("pyproject.toml"), "").expect("write pyproject.toml");
-        std::fs::write(dir.path().join("utils.py"), "def helper(): pass\n").expect("write");
-        std::fs::write(
-            dir.path().join("test_foo.py"),
-            "from utils import helper\n@test\ndef test_foo(): pass\n",
-        )
-        .expect("write");
-        std::fs::write(
-            dir.path().join("test_isolated.py"),
-            "@test\ndef test_isolated(): pass\n",
-        )
-        .expect("write");
-        assert!(run_graph(Some(dir.path()), &[], true, false, None).is_ok());
-    }
-
-    fn init_git_repo(dir: &tempfile::TempDir) {
-        fn run(dir: &tempfile::TempDir, args: &[&str]) {
-            let status = std::process::Command::new("git")
-                .args(args)
-                .current_dir(dir.path())
-                .status()
-                .expect("run git");
-            assert!(status.success(), "git {:?} failed", args);
-        }
-
-        run(dir, &["init"]);
-        run(dir, &["config", "user.email", "tryke@example.com"]);
-        run(dir, &["config", "user.name", "Tryke Tests"]);
-        run(dir, &["config", "commit.gpgsign", "false"]);
-    }
-
-    fn git_run(dir: &tempfile::TempDir, args: &[&str]) {
-        let status = std::process::Command::new("git")
-            .args(args)
-            .current_dir(dir.path())
-            .status()
-            .expect("run git");
-        assert!(status.success(), "git {:?} failed", args);
-    }
-
-    /// Seed a git repo with an initial commit containing `pyproject.toml`
-    /// and an optional set of extra files, so `git diff --name-only HEAD`
-    /// has a baseline to compare against.
-    fn seed_git_repo(dir: &tempfile::TempDir, files: &[(&str, &str)]) {
-        std::fs::write(dir.path().join("pyproject.toml"), "").expect("write pyproject.toml");
-        for &(name, content) in files {
-            if let Some(parent) = std::path::Path::new(name).parent()
-                && !parent.as_os_str().is_empty()
-            {
-                std::fs::create_dir_all(dir.path().join(parent)).expect("mkdir");
-            }
-            std::fs::write(dir.path().join(name), content).expect("write file");
-        }
-        init_git_repo(dir);
-        git_run(dir, &["add", "."]);
-        git_run(dir, &["commit", "-m", "initial"]);
-    }
-
-    #[test]
-    fn git_changed_files_includes_untracked() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo(&dir, &[("tracked.py", "def helper(): pass\n")]);
-
-        std::fs::write(
-            dir.path().join("test_new.py"),
-            "@test\ndef test_new(): pass\n",
-        )
-        .expect("write untracked file");
-
-        let changed = git_changed_files(dir.path()).expect("git changed files");
-        assert!(changed.contains(&dir.path().join("test_new.py")));
-    }
-
-    #[test]
-    fn git_changed_files_includes_tracked_modifications() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo(&dir, &[("lib.py", "x = 1\n")]);
-
-        std::fs::write(dir.path().join("lib.py"), "x = 2\n").expect("modify tracked file");
-
-        let changed = git_changed_files(dir.path()).expect("git changed files");
-        assert!(changed.contains(&dir.path().join("lib.py")));
-    }
-
-    #[test]
-    fn git_changed_files_returns_empty_when_clean() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo(&dir, &[("lib.py", "x = 1\n")]);
-
-        let changed = git_changed_files(dir.path()).expect("git changed files");
-        assert!(changed.is_empty());
-    }
-
-    #[test]
-    fn git_changed_files_includes_staged_files() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo(&dir, &[]);
-
-        std::fs::write(dir.path().join("new.py"), "y = 1\n").expect("write new file");
-        git_run(&dir, &["add", "new.py"]);
-
-        let changed = git_changed_files(dir.path()).expect("git changed files");
-        assert!(changed.contains(&dir.path().join("new.py")));
-    }
-
-    #[test]
-    fn git_changed_files_deduplicates_tracked_and_untracked() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo(&dir, &[]);
-
-        // Stage a new file, then modify it again so it appears in both
-        // `git diff --name-only HEAD` (staged) and `git ls-files --others`
-        // would not list it since it's tracked. But the staged version differs
-        // from HEAD, and the working copy differs from the index, so it
-        // appears in `git diff --name-only HEAD` (which covers both).
-        // This verifies our BTreeSet dedup doesn't produce duplicates.
-        std::fs::write(dir.path().join("dup.py"), "v1\n").expect("write");
-        git_run(&dir, &["add", "dup.py"]);
-        std::fs::write(dir.path().join("dup.py"), "v2\n").expect("modify");
-
-        let changed = git_changed_files(dir.path()).expect("git changed files");
-        let count = changed
-            .iter()
-            .filter(|p| *p == &dir.path().join("dup.py"))
-            .count();
-        assert_eq!(count, 1, "file should appear exactly once");
-    }
-
-    #[test]
-    fn git_changed_files_handles_paths_with_spaces() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo(&dir, &[]);
-
-        let subdir = dir.path().join("my tests");
-        std::fs::create_dir_all(&subdir).expect("mkdir");
-        std::fs::write(subdir.join("test_space.py"), "pass\n").expect("write");
-
-        let changed = git_changed_files(dir.path()).expect("git changed files");
-        assert!(
-            changed.contains(&subdir.join("test_space.py")),
-            "should detect files in directories with spaces: {changed:?}"
-        );
-    }
-
-    #[test]
-    fn git_changed_files_includes_deleted_files() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo(&dir, &[("to_delete.py", "pass\n")]);
-
-        std::fs::remove_file(dir.path().join("to_delete.py")).expect("delete file");
-
-        let changed = git_changed_files(dir.path()).expect("git changed files");
-        assert!(
-            changed.contains(&dir.path().join("to_delete.py")),
-            "deleted files should appear in changed list: {changed:?}"
-        );
-    }
-
-    #[test]
-    fn git_changed_files_returns_sorted() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo(&dir, &[]);
-
-        std::fs::write(dir.path().join("z.py"), "pass\n").expect("write");
-        std::fs::write(dir.path().join("a.py"), "pass\n").expect("write");
-        std::fs::write(dir.path().join("m.py"), "pass\n").expect("write");
-
-        let changed = git_changed_files(dir.path()).expect("git changed files");
-        let is_sorted = changed.windows(2).all(|w| w[0] <= w[1]);
-        assert!(is_sorted, "results should be sorted: {changed:?}");
-    }
-
-    #[tokio::test]
-    async fn run_changed_test_without_git_runs_all() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("pyproject.toml"), "").expect("write pyproject.toml");
-        let mut reporter = TextReporter::new();
-        // non-git directory → git_changed_files returns None → discover_tests runs all (0 here)
-        let tests = discover_tests(dir.path(), true, None, &[]).tests;
-        assert!(
-            run_tests(&mut reporter, dir.path(), tests, None, None, None, None)
-                .await
-                .is_ok()
-        );
     }
 
     #[test]
@@ -1381,312 +560,28 @@ mod tests {
         ));
     }
 
-    #[tokio::test]
-    async fn integration_python_worker_runs_tests() {
-        let workspace_root = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .join("../..")
-            .canonicalize()
-            .expect("workspace root");
-        let python = test_python_bin();
-        tryke_runner::check_python_version(&python, &workspace_root)
-            .expect("Python version check (from pyproject.toml requires-python)");
-
-        let python_dir = workspace_root.join("python");
-
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::write(dir.path().join("pyproject.toml"), "").expect("write pyproject.toml");
-        std::fs::write(
-            dir.path().join("test_example.py"),
-            "\
-from tryke import test, expect
-
-@test
-def test_passing():
-    expect(1 + 1).to_equal(2)
-
-@test
-def test_failing():
-    expect(1 + 1).to_equal(3)
-",
-        )
-        .expect("write test file");
-
-        let tests = discover_tests(dir.path(), false, None, &[]).tests;
-        assert_eq!(tests.len(), 2);
-
-        let pool = WorkerPool::with_python_path(
-            1,
-            &test_python_bin(),
-            dir.path(),
-            &[dir.path().to_path_buf(), python_dir],
-        );
-        pool.warm().await;
-        let mut results: Vec<_> = pool.run(tests).collect().await;
-        results.sort_by(|a, b| a.test.name.cmp(&b.test.name));
-
-        assert_eq!(results.len(), 2);
-        assert!(
-            matches!(results[0].outcome, TestOutcome::Failed { .. }),
-            "test_failing should fail, got {:?}",
-            results[0].outcome
-        );
-        assert!(
-            matches!(results[1].outcome, TestOutcome::Passed),
-            "test_passing should pass, got {:?}",
-            results[1].outcome
-        );
-        for r in &results {
-            assert!(
-                !matches!(r.outcome, TestOutcome::Error { .. }),
-                "unexpected worker error: {:?}",
-                r.outcome
-            );
-        }
-
-        pool.shutdown();
+    #[test]
+    fn graph_subcommand_parsed() {
+        let cli = Cli::try_parse_from(["tryke", "graph"]).unwrap();
+        assert!(matches!(cli.command, Commands::Graph { .. }));
     }
 
-    // --- Branch mode tests ---
+    #[test]
+    fn graph_subcommand_changed_parsed() {
+        let cli = Cli::try_parse_from(["tryke", "graph", "--changed"]).unwrap();
+        assert!(matches!(cli.command, Commands::Graph { changed: true, .. }));
+    }
 
-    /// Seed a git repo with an explicit "main" branch name for branch-mode tests.
-    fn seed_git_repo_with_main(dir: &tempfile::TempDir, files: &[(&str, &str)]) {
-        std::fs::write(dir.path().join("pyproject.toml"), "").expect("write pyproject.toml");
-        for &(name, content) in files {
-            if let Some(parent) = std::path::Path::new(name).parent()
-                && !parent.as_os_str().is_empty()
-            {
-                std::fs::create_dir_all(dir.path().join(parent)).expect("mkdir");
+    #[test]
+    fn graph_subcommand_connected_only_parsed() {
+        let cli = Cli::try_parse_from(["tryke", "graph", "--connected-only"]).unwrap();
+        assert!(matches!(
+            cli.command,
+            Commands::Graph {
+                connected_only: true,
+                ..
             }
-            std::fs::write(dir.path().join(name), content).expect("write file");
-        }
-        init_git_repo(dir);
-        git_run(dir, &["checkout", "-b", "main"]);
-        git_run(dir, &["add", "."]);
-        git_run(dir, &["commit", "-m", "initial"]);
-    }
-
-    #[test]
-    fn git_branch_changed_files_returns_branch_diff() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo_with_main(&dir, &[("base.py", "x = 1\n")]);
-
-        git_run(&dir, &["checkout", "-b", "feature"]);
-        std::fs::write(dir.path().join("feature.py"), "y = 2\n").expect("write");
-        git_run(&dir, &["add", "feature.py"]);
-        git_run(&dir, &["commit", "-m", "feature commit"]);
-
-        let changed =
-            git_branch_changed_files(dir.path(), "main").expect("git branch changed files");
-        assert!(
-            changed.contains(&dir.path().join("feature.py")),
-            "branch file should appear: {changed:?}"
-        );
-        assert!(
-            !changed.contains(&dir.path().join("base.py")),
-            "base file should not appear: {changed:?}"
-        );
-    }
-
-    #[test]
-    fn git_branch_changed_files_includes_untracked() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo_with_main(&dir, &[("base.py", "x = 1\n")]);
-
-        git_run(&dir, &["checkout", "-b", "feature"]);
-        std::fs::write(dir.path().join("feature.py"), "y = 2\n").expect("write");
-        git_run(&dir, &["add", "feature.py"]);
-        git_run(&dir, &["commit", "-m", "feature commit"]);
-
-        std::fs::write(dir.path().join("untracked.py"), "z = 3\n").expect("write");
-
-        let changed =
-            git_branch_changed_files(dir.path(), "main").expect("git branch changed files");
-        assert!(
-            changed.contains(&dir.path().join("untracked.py")),
-            "untracked file should appear: {changed:?}"
-        );
-    }
-
-    #[test]
-    fn git_branch_changed_files_nonexistent_branch_returns_none() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo(&dir, &[("base.py", "x = 1\n")]);
-
-        let result = git_branch_changed_files(dir.path(), "nonexistent-branch-xyz");
-        assert!(result.is_none(), "nonexistent branch should return None");
-    }
-
-    #[test]
-    fn git_branch_changed_files_excludes_main_only_changes() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo_with_main(&dir, &[("base.py", "x = 1\n")]);
-
-        git_run(&dir, &["checkout", "-b", "feature"]);
-        std::fs::write(dir.path().join("feature.py"), "y = 2\n").expect("write");
-        git_run(&dir, &["add", "feature.py"]);
-        git_run(&dir, &["commit", "-m", "feature commit"]);
-
-        // Switch back to main and add a change there
-        git_run(&dir, &["checkout", "main"]);
-        std::fs::write(dir.path().join("main_only.py"), "z = 3\n").expect("write");
-        git_run(&dir, &["add", "main_only.py"]);
-        git_run(&dir, &["commit", "-m", "main-only commit"]);
-
-        // Switch back to feature branch
-        git_run(&dir, &["checkout", "feature"]);
-
-        let changed =
-            git_branch_changed_files(dir.path(), "main").expect("git branch changed files");
-        assert!(
-            !changed.contains(&dir.path().join("main_only.py")),
-            "main-only changes should not appear: {changed:?}"
-        );
-        assert!(
-            changed.contains(&dir.path().join("feature.py")),
-            "feature changes should appear: {changed:?}"
-        );
-    }
-
-    #[test]
-    fn discover_tests_with_base_branch() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo_with_main(
-            &dir,
-            &[(
-                "test_base.py",
-                "from tryke import test\n\n@test\ndef test_base(): pass\n",
-            )],
-        );
-
-        git_run(&dir, &["checkout", "-b", "feature"]);
-        std::fs::write(
-            dir.path().join("test_feature.py"),
-            "from tryke import test\n\n@test\ndef test_feature(): pass\n",
-        )
-        .expect("write");
-        git_run(&dir, &["add", "test_feature.py"]);
-        git_run(&dir, &["commit", "-m", "add feature test"]);
-
-        let discovered = discover_tests(dir.path(), true, Some("main"), &[]);
-        assert!(
-            discovered.tests.iter().any(|t| t.name == "test_feature"),
-            "should find the branch's test: {:?}",
-            discovered.tests.iter().map(|t| &t.name).collect::<Vec<_>>()
-        );
-    }
-
-    // --- Changed-first tests ---
-
-    #[test]
-    fn discover_tests_changed_first_partitions_correctly() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo(
-            &dir,
-            &[
-                (
-                    "test_a.py",
-                    "from tryke import test\n\n@test\ndef test_a(): pass\n",
-                ),
-                (
-                    "test_b.py",
-                    "from tryke import test\n\n@test\ndef test_b(): pass\n",
-                ),
-            ],
-        );
-
-        // Modify test_a.py so it counts as "changed"
-        std::fs::write(
-            dir.path().join("test_a.py"),
-            "from tryke import test\n\n@test\ndef test_a(): assert True\n",
-        )
-        .expect("write");
-
-        let discovered = discover_tests_changed_first(dir.path(), None, &[]);
-        let names: Vec<&str> = discovered.tests.iter().map(|t| t.name.as_str()).collect();
-
-        assert!(
-            discovered.changed_prefix_len.is_some(),
-            "changed_prefix_len should be set"
-        );
-        let prefix_len = discovered.changed_prefix_len.expect("set");
-        assert!(prefix_len > 0, "at least one changed test");
-        // Changed test(s) should be at the front
-        let changed_names: Vec<&str> = names[..prefix_len].to_vec();
-        assert!(
-            changed_names.contains(&"test_a"),
-            "test_a should be in the changed prefix: {names:?}"
-        );
-        // All tests should still be present
-        assert!(
-            names.contains(&"test_b"),
-            "test_b should still be present: {names:?}"
-        );
-    }
-
-    #[test]
-    fn discover_tests_changed_first_no_changes() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo(
-            &dir,
-            &[(
-                "test_a.py",
-                "from tryke import test\n\n@test\ndef test_a(): pass\n",
-            )],
-        );
-
-        let discovered = discover_tests_changed_first(dir.path(), None, &[]);
-        assert!(
-            discovered.changed_prefix_len.is_none(),
-            "changed_prefix_len should be None when no changes"
-        );
-        assert!(
-            !discovered.tests.is_empty(),
-            "all tests should still be returned"
-        );
-    }
-
-    #[test]
-    fn discover_tests_changed_first_with_base_branch() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        seed_git_repo_with_main(
-            &dir,
-            &[
-                (
-                    "test_a.py",
-                    "from tryke import test\n\n@test\ndef test_a(): pass\n",
-                ),
-                (
-                    "test_b.py",
-                    "from tryke import test\n\n@test\ndef test_b(): pass\n",
-                ),
-            ],
-        );
-
-        git_run(&dir, &["checkout", "-b", "feature"]);
-        std::fs::write(
-            dir.path().join("test_c.py"),
-            "from tryke import test\n\n@test\ndef test_c(): pass\n",
-        )
-        .expect("write");
-        git_run(&dir, &["add", "test_c.py"]);
-        git_run(&dir, &["commit", "-m", "add test_c"]);
-
-        let discovered = discover_tests_changed_first(dir.path(), Some("main"), &[]);
-        let names: Vec<&str> = discovered.tests.iter().map(|t| t.name.as_str()).collect();
-
-        assert!(
-            discovered.changed_prefix_len.is_some(),
-            "changed_prefix_len should be set"
-        );
-        let prefix_len = discovered.changed_prefix_len.expect("set");
-        // test_c should be in the changed prefix
-        let changed_names: Vec<&str> = names[..prefix_len].to_vec();
-        assert!(
-            changed_names.contains(&"test_c"),
-            "test_c should be in the changed prefix: {names:?}"
-        );
-        // All 3 tests should be present
-        assert_eq!(names.len(), 3, "all 3 tests should be present: {names:?}");
+        ));
     }
 
     // --- CLI parsing tests for new flags ---
